+++ Source/Fractorium/FractoriumXformsVariations.cpp
@@ -50,7 +50,7 @@ void FractoriumEmberController<T>::Filter(const QString& text)
 
 	tree->setUpdatesEnabled(false);
 
-	for (uint i = 0; i < uint(tree->topLevelItemCount()); i++)
+	for (int i = 0; i < tree->topLevelItemCount(); i++)
 	{
 		if (auto item = dynamic_cast<VariationTreeWidgetItem*>(tree->topLevelItem(i)))
 		{
@@ -181,14 +181,14 @@ void FractoriumEmberController<T>::ClearVariationsTree()
 {
 	QTreeWidget* tree = m_Fractorium->ui.VariationsTree;
 
-	for (uint i = 0; i < tree->topLevelItemCount(); i++)
+	for (int i = 0; i < tree->topLevelItemCount(); i++)
 	{
 		QTreeWidgetItem* item = tree->topLevelItem(i);
 		auto* spinBox = dynamic_cast<VariationTreeDoubleSpinBox*>(tree->itemWidget(item, 1));
 
 		spinBox->SetValueStealth(0);
 
-		for (uint j = 0; j < item->childCount(); j++)//Iterate through all of the children, which will be the params.
+		for (int j = 0; j < item->childCount(); j++)//Iterate through all of the children, which will be the params.
 		{
 			if ((spinBox = dynamic_cast<VariationTreeDoubleSpinBox*>(tree->itemWidget(item->child(j), 1))))//Cast the child widget to the VariationTreeDoubleSpinBox type.
 				spinBox->SetValueStealth(0);
@@ -301,7 +301,7 @@ void FractoriumEmberController<T>::FillVariationTreeWithXform(Xform<T>* xform)
 	tree->blockSignals(true);
 	m_Fractorium->Filter();
 
-	for (uint i = 0; i < tree->topLevelItemCount(); i++)
+	for (int i = 0; i < tree->topLevelItemCount(); i++)
 	{
 		auto item = dynamic_cast<VariationTreeWidgetItem*>(tree->topLevelItem(i));
 		auto var = xform->GetVariationById(item->Id());//See if this variation in the tree was contained in the xform.
@@ -317,7 +317,7 @@ void FractoriumEmberController<T>::FillVariationTreeWithXform(Xform<T>* xform)
 			//item->setBackgroundColor(0, var ? Qt::darkGray : Qt::lightGray);//Ensure background is always white if the value goes to zero, else gray if var present.
 			item->setBackgroundColor(0, var ? QColor(200, 200, 200) : QColor(255, 255, 255));//Ensure background is always white if the value goes to zero, else gray if var present.
 
-			for (uint j = 0; j < item->childCount(); j++)//Iterate through all of the children, which will be the params if it was a parametric variation.
+			for (int j = 0; j < item->childCount(); j++)//Iterate through all of the children, which will be the params if it was a parametric variation.
 			{
 				T* param = nullptr;
 				auto childItem = item->child(j);//Get the child.
