+++ Source/Fractorium/GLEmberController.cpp
@@ -252,13 +252,13 @@ void GLEmberController<T>::QueryMatrices(bool print)
 
 		if (print)
 		{
-			for (int i = 0; i < 4; i++)
+			for (size_t i = 0; i < 4; i++)
 				qDebug() << "Viewport[" << i << "] = " << m_Viewport[i] << endl;
 
-			for (int i = 0; i < 16; i++)
+			for (size_t i = 0; i < 16; i++)
 				qDebug() << "Modelview[" << i << "] = " << glm::value_ptr(m_Modelview)[i] << endl;
 
-			for (int i = 0; i < 16; i++)
+			for (size_t i = 0; i < 16; i++)
 				qDebug() << "Projection[" << i << "] = " << glm::value_ptr(m_Projection)[i] << endl;
 		}
 	}
